courseapisv2/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
pillow==11.0.0
PyMySQL==1.1.1
sqlparse==0.5.1
tzdata==2024.2
